requirements.txt
psutil
pyautogui
opencv-python
pycryptodome
requests
pillow
pywin32
browser-history
pywifi
pyinstaller 